requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
cloudinary==1.42.2
python-dotenv==1.0.1
requests==2.32.3
